splited WishList
