+++ src/components/WishListView/WishListView.test.js
@@ -4,7 +4,7 @@ import { render } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import WishListTable from './WishListTable';
+import WishListTable from './WishListView';
 
 const mockStore = configureStore([]);
 const products = [
